+++ connectlib/algorithms/pytorch/torch_single_organization_algo.py
@@ -25,13 +25,11 @@ class TorchSingleOrganizationAlgo(TorchAlgo):
         - calls the :py:func:`~connectlib.algorithms.pytorch.TorchSingleOrganizationAlgo._local_train` method to do the
           local training
 
-    The ``predict`` method calls the
-    :py:func:`~connectlib.algorithms.pytorch.TorchSingleOrganizationAlgo._local_predict`
-    method to generate the predictions.
+    The ``predict`` method generates the predictions.
 
-    The child class must implement the
+    The child class can override the
     :py:func:`~connectlib.algorithms.pytorch.TorchSingleOrganizationAlgo._local_train`
-    and :py:func:`~connectlib.algorithms.pytorch.TorchSingleOrganizationAlgo._local_predict` methods, and can override
+    and :py:func:`~connectlib.algorithms.pytorch.TorchSingleOrganizationAlgo.predict` methods, or
     other methods if necessary.
 
     To add a custom parameter to the ``__init__``of the class, also add it to the call to ``super().__init__```
@@ -55,17 +53,20 @@ class TorchSingleOrganizationAlgo(TorchAlgo):
                             num_updates=100,
                             batch_size=32,
                         ),
+                        dataset=MyDataset,
                         my_custom_extra_parameter=my_custom_extra_parameter,
                     )
                 def _local_train(
                     self,
-                    x: Any,
-                    y: Any,
+                    train_dataset: torch.utils.data.Dataset,
                 ):
-                    # for each update
-                    for batch_index in self._index_generator:
-                        # get minibatch
-                        x_batch, y_batch = x[batch_index], y[batch_index]
+                    # Create torch dataloader
+                    # ``train_dataset = self._dataset(x=x, y=y, is_inference=False)`` is executed prior the execution
+                    # of this function
+                    train_data_loader = torch.utils.data.DataLoader(train_dataset, batch_sampler=self._index_generator)
+
+                    for x_batch, y_batch in train_data_loader:
+
                         # Forward pass
                         y_pred = self._model(x_batch)
                         # Compute Loss
@@ -79,22 +80,6 @@ class TorchSingleOrganizationAlgo(TorchAlgo):
                         if self._scheduler is not None:
                             self._scheduler.step()
 
-                def _local_predict(self, x: Any) -> Any:
-                    with torch.inference_mode():
-                        y = self._model(x)
-                    return y
-
-    The algo needs to get the number of samples in the dataset from the x sent by the opener.
-    By default, it uses len(x). If that is not the proper way of getting the number of samples,
-    override the ``_get_len_from_x`` function
-
-    Example:
-
-        .. code-block:: python
-
-            def _get_len_from_x(self, x):
-                return len(x)
-
     As development tools, the ``train`` and ``predict`` method comes with a default argument : ``_skip``.
     If ``_skip`` is set to ``True``, only the function will be executed and not all the code related to Connect.
     This allows to quickly debug code and use the defined algorithm as is.
@@ -106,6 +91,7 @@ class TorchSingleOrganizationAlgo(TorchAlgo):
         criterion: torch.nn.modules.loss._Loss,
         optimizer: torch.optim.Optimizer,
         index_generator: BaseIndexGenerator,
+        dataset: torch.utils.data.Dataset,
         scheduler: Optional[torch.optim.lr_scheduler._LRScheduler] = None,
         use_gpu: bool = True,
         *args,
@@ -126,6 +112,15 @@ class TorchSingleOrganizationAlgo(TorchAlgo):
                 during training in this way: ``x[batch_index], y[batch_index]``.
                 If overridden, the generator class must be defined either as part of a package or in a different file
                 than the one from which the ``execute_experiment`` function is called.
+                This generator is used as stateful ``batch_sampler`` of the data loader created from the given
+                ``dataset``
+            dataset (torch.utils.data.Dataset): an instantiable dataset class whose ``__init__`` arguments are
+                ``x``, ``y`` and ``is_inference``. The torch datasets used for both training and inference will be
+                instantiate from it prior to the ``_local_train`` execution and within the ``predict`` method.
+                The ``__getitem__`` methods of those generated datasets must return both ``x`` (training data) and y
+                (target values) when ``is_inference`` is set to ``False`` and only ``x`` (testing data) when
+                ``is_inference`` is set to True.
+                This behavior can be changed by re-writing the `_local_train` or `predict` methods.
             scheduler (torch.optim.lr_scheduler._LRScheduler, Optional): A torch scheduler that will be called at every
                 batch. If None, no scheduler will be used. Defaults to None.
             use_gpu (bool): Whether to use the GPUs if they are available. Defaults to True.
@@ -135,6 +130,7 @@ class TorchSingleOrganizationAlgo(TorchAlgo):
             criterion=criterion,
             optimizer=optimizer,
             index_generator=index_generator,
+            dataset=dataset,
             scheduler=scheduler,
             use_gpu=use_gpu,
             *args,
@@ -170,20 +166,23 @@ class TorchSingleOrganizationAlgo(TorchAlgo):
             shared state and this return is only for consistency.
         """
 
+        # Create torch dataset
+        train_dataset = self._dataset(x=x, y=y, is_inference=False)
+
         # Instantiate the index_generator
         if self._index_generator.n_samples is None:
-            self._index_generator.n_samples = self._get_len_from_x(x)
+            self._index_generator.n_samples = len(train_dataset)
 
         self._index_generator.reset_counter()
+
+        # Create torch dataloader
+        self._torch_loader = torch.utils.data.DataLoader(train_dataset, batch_sampler=self._index_generator)
 
         # Train mode for torch model
         self._model.train()
 
         # Train the model
-        self._local_train(
-            x=x,
-            y=y,
-        )
+        self._local_train(train_dataset)
 
         self._index_generator.check_num_updates()