+++ docs/conf.py
@@ -184,6 +184,7 @@ nitpick_ignore = [
     ("py:class", "torch.optim.optimizer.Optimizer"),
     ("py:class", "torch.optim.lr_scheduler._LRScheduler"),
     ("py:class", "torch.device"),
+    ("py:class", "torch.utils.data.dataset.Dataset"),
     ("py:class", "substra.sdk.schemas.Permissions"),
     ("py:class", "substra.Client"),
     ("py:class", "substra.sdk.client.Client"),